requirements.txt
Flask==1.1.2
numpy==1.20.3
opencv_python==4.5.5.64
Pillow==9.2.0
tensorflow==2.9.1
tensorflow_gpu==2.9.1
